requirements.txt
# Core Dependencies
numpy>=1.24.0
torch>=2.0.0
gymnasium>=0.29.0

# Crafter Environment
# crafter>=1.8.0  # temporarily commented out for initial dependency install; install separately later

# Deep Learning & RL
stable-baselines3>=2.0.0
tensorboard>=2.13.0

# LLM Dependencies
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.40.0
peft>=0.4.0
datasets>=2.14.0
sentencepiece>=0.1.99

# OpenAI API (for Helper)
openai>=1.0.0
anthropic>=0.3.0

# Utilities
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
wandb>=0.15.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
ipykernel>=6.25.0
jupyter>=1.0.0
